Json contains information about frame duration.
* Extended JSONs containing animation informations with frameDuration.
* Adjusted code changes in Animation class.
* Replaced new keywoard that was allocating shared pointer with Texture with std::make_shared

+++ src/core/include/Animation.hpp
@@ -31,7 +31,7 @@ struct FrameData
   int width;
   int height;
 
-  float displayTime;
+  double displayTime;
 };
 
 class Animation
